requirements.txt
RPi.GPIO
RPi-GPIO-Rotary
miniaudio
fastapi
uvicorn
python-multipart
lupa  
pycairo 
numpy 